backend/requirements.txt
Django>=5.0,<6.0
psycopg2-binary>=2.9
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
drf-spectacular>=0.27
